refactor: 패키지 네이밍 변경 및 클래스 이동
