+++ core-application/src/main/java/gohigher/port/in/UserCommandPort.java
@@ -1,7 +1,7 @@
 package gohigher.port.in;
 
 import gohigher.user.User;
-import gohigher.user.oauth2.Provider;
+import gohigher.user.auth.Provider;
 
 public interface UserCommandPort {
