+++ core-domain/src/main/java/gohigher/user/User.java
@@ -1,6 +1,6 @@
 package gohigher.user;
 
-import gohigher.user.oauth2.Provider;
+import gohigher.user.auth.Provider;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -8,10 +8,10 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public class User {
 
-	private Long id;
 	private final String email;
 	private final Role role;
 	private final Provider provider;
+	private Long id;
 
 	public User(Long id, String email, Role role, Provider provider) {
 		this.id = id;
